GetRemainingExposureTime no longer rounds to nearest second
[git-p4: depot-paths = "//depot/ACG/Smith/Firmware/": change = 97942]


Former-commit-id: 91eb1da4ba1545862aa9aafd6520a4b161dc3263

+++ C++/include/PrintEngine.h
@@ -64,7 +64,7 @@ public:
     void CancelPrint();
     double GetExposureTimeSec();
     char GetSeparationCommand();
-    int GetRemainingExposureTimeSec();
+    double GetRemainingExposureTimeSec();
     bool DoorIsOpen();
     I2C_Device* GetMotorBoard() { return _pMotor; }
     void ShowImage();
